requirements.txt
git+git://github.com/bobylevd/Twitch-Python@master
redis
discord.py
